feat: adding delete usecase and refactoring DDD
