+++ src/src/Application/UseCase/Car/CreateCar/CreateCarUseCase.php
@@ -1,22 +1,21 @@
 <?php
 
-namespace App\Application\UseCase;
+namespace App\Application\UseCase\Car\CreateCar;
 
 use App\DTO\CarDTO;
 use App\Mapper\CarMapper;
-use Doctrine\ORM\EntityManagerInterface;
+use App\Domain\Repository\CarRepositoryInterface;
 
 class CreateCarUseCase implements CreateCarUseCaseInterface
 {
   public function __construct(
-    private readonly EntityManagerInterface $entityManager,
+    private readonly CarRepositoryInterface $carRepository,
     private readonly CarMapper $carMapper
   ) {}
 
   public function execute(CarDTO $dto): void
   {
     $car = $this->carMapper->mapToEntity($dto);
-    $this->entityManager->persist($car);
-    $this->entityManager->flush();
+    $this->carRepository->save($car);
   }
 }
