+++ src/src/Application/UseCase/Car/DeleteCar/DeleteCarUseCaseInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Application\UseCase\Car\DeleteCar;
+
+use App\DTO\CarDTO;
+
+interface DeleteCarUseCaseInterface
+{
+	public function execute(CarDTO $dto): void;
+}
