+++ src/src/Domain/Repository/CarRepositoryInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Domain\Repository;
+
+use App\Domain\Entity\Car;
+
+interface CarRepositoryInterface
+{
+  public function save(Car $car): void;
+  public function remove(Car $car): void;
+}
