+++ src/src/Infrastructure/Repository/CarRepository.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace App\Infrastructure\Repository;
+
+use App\Domain\Repository\CarRepositoryInterface;
+use App\Domain\Entity\Car;
+use Doctrine\ORM\EntityManagerInterface;
+
+class CarRepository implements CarRepositoryInterface
+{
+	public function __construct(private EntityManagerInterface $entityManager) {}
+
+	public function save(Car $car): void
+	{
+		$this->entityManager->persist($car);
+		$this->entityManager->flush();
+	}
+
+	public function remove(Car $car): void
+	{
+		$this->entityManager->remove($car);
+		$this->entityManager->flush();
+	}
+}
